+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [Unreleased]
 ### Changed
+- (RE-14611) Change msi sign method to use the gcp msi signer
 - (RE-13764) Get rvm ruby version from environment variable instead of hard coded in remote
   bundle install method.